Andrus Jakobson: Changed file name and messed around with the image.
