+++ AndroidStudioProjects/projectNAM2/app/src/main/java/com/example/andrus/projectnam/MainActivity.java
@@ -5,7 +5,7 @@ import android.support.v4.app.FragmentTransaction;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 
-import com.example.andrus.projectnam.MoodGrid.MoodGridFragment;
+import com.example.andrus.projectnam.moodgrid.MoodGridFragment;
 
 public class MainActivity extends AppCompatActivity {
