--- AndroidStudioProjects/projectNAM2/app/src/main/java/com/example/andrus/projectnam/Models/Mood.java
@@ -1,7 +1,0 @@
-package com.example.andrus.projectnam.Models;
-
-public class Mood {
-    public int categoryId;
-    public String categoryName;
-    public String CategoryLogo;
-}
