--- AndroidStudioProjects/projectNAM2/app/src/main/java/com/example/andrus/projectnam/MoodDetail/MoodDetailInterface.java
@@ -1,6 +1,0 @@
-package com.example.andrus.projectnam.MoodDetail;
-
-
-public interface MoodDetailInterface {
-
-}
