--- AndroidStudioProjects/projectNAM2/app/src/main/java/com/example/andrus/projectnam/MoodDetail/MoodDetailPresenter.java
@@ -1,5 +1,0 @@
-package com.example.andrus.projectnam.MoodDetail;
-
-public class MoodDetailPresenter {
-
-}
