+++ AndroidStudioProjects/projectNAM2/app/src/main/java/com/example/andrus/projectnam/models/Mood.java
@@ -1,0 +1,10 @@
+package com.example.andrus.projectnam.models;
+
+import android.media.Image;
+import android.util.Base64;
+
+public class Mood {
+    public int categoryId;
+    public String categoryName;
+    public Base64 CategoryLogo;
+}
