+++ AndroidStudioProjects/projectNAM2/app/src/main/java/com/example/andrus/projectnam/models/Moods.java
@@ -1,4 +1,4 @@
-package com.example.andrus.projectnam.Models;
+package com.example.andrus.projectnam.models;
 
 
 import java.util.List;
