+++ AndroidStudioProjects/projectNAM2/app/src/main/java/com/example/andrus/projectnam/mooddetails/MoodDetailInterface.java
@@ -1,0 +1,6 @@
+package com.example.andrus.projectnam.mooddetails;
+
+
+public interface MoodDetailInterface {
+
+}
