+++ AndroidStudioProjects/projectNAM2/app/src/main/java/com/example/andrus/projectnam/mooddetails/MoodDetailPresenter.java
@@ -1,0 +1,5 @@
+package com.example.andrus.projectnam.mooddetails;
+
+public class MoodDetailPresenter {
+
+}
