+++ AndroidStudioProjects/projectNAM2/app/src/main/java/com/example/andrus/projectnam/moodgrid/MoodGridInterface.java
@@ -1,7 +1,6 @@
-package com.example.andrus.projectnam.MoodGrid;
+package com.example.andrus.projectnam.moodgrid;
 
-import com.example.andrus.projectnam.Models.Mood;
-import com.example.andrus.projectnam.Models.Moods;
+import com.example.andrus.projectnam.models.Mood;
 
 import java.util.List;
