+++ AndroidStudioProjects/projectNAM2/app/src/main/java/com/example/andrus/projectnam/moodgrid/MoodGridRequester.java
@@ -1,7 +1,6 @@
-package com.example.andrus.projectnam.MoodGrid;
+package com.example.andrus.projectnam.moodgrid;
 
-import com.example.andrus.projectnam.Models.Mood;
-import com.example.andrus.projectnam.Models.Moods;
+import com.example.andrus.projectnam.models.Mood;
 import com.example.andrus.projectnam.util.http.APIClient;
 
 import java.util.List;
