+++ AndroidStudioProjects/projectNAM2/app/src/main/java/com/example/andrus/projectnam/util/http/APIService.java
@@ -1,8 +1,7 @@
 package com.example.andrus.projectnam.util.http;
 
 
-import com.example.andrus.projectnam.Models.Mood;
-import com.example.andrus.projectnam.Models.Moods;
+import com.example.andrus.projectnam.models.Mood;
 
 import java.util.List;
 
@@ -12,4 +11,6 @@ import retrofit2.http.GET;
 public interface APIService {
     @GET("5a29408e2e00000f1ea09b40")
     Call<List<Mood>> getMoodGridIconAndText();
+
+    
 }
