+++ AndroidStudioProjects/projectNAM2/app/src/main/res/layout/adapter_main_screen.xml
@@ -8,7 +8,6 @@
         android:layout_width="100dp"
         android:layout_height="100dp"
         android:layout_centerHorizontal="true"
-        android:background="#000"
         android:contentDescription="@string/moodimage" />
 
     <TextView
